backend/wearable_backend/app/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
openai
chromadb